fix: add a correction in the way the Context is created; all ESLint errors have been resolved, and the issue with the title bar overlapping has also been fixed.
