+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/mem-gui.svg" />
+    <link rel="icon" type="image/svg+xml" href="/images/logo-white.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>MemGUI</title>
   </head>
