+++ src/ui/components/KeyList.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-hooks/exhaustive-deps */
 import {
   ArrowPathIcon,
   DocumentMagnifyingGlassIcon,
