+++ src/ui/contexts/MenuContext.tsx
@@ -1,7 +1,6 @@
-import { ReactNode, useState } from "react";
-import React, { createContext } from "react";
+import { createContext } from "react";
 
-interface MenuContextType {
+export interface MenuContextType {
   openMenu: () => void;
   closeMenu: () => void;
   menuIsOpen: boolean;
@@ -10,21 +9,3 @@ interface MenuContextType {
 export const MenuContext = createContext<MenuContextType | undefined>(
   undefined
 );
-
-export const MenuProvider = ({ children }: { children: ReactNode }) => {
-  const [menuIsOpen, setMenuIsOpen] = useState(false);
-
-  const openMenu = () => {
-    setMenuIsOpen(true);
-  };
-
-  const closeMenu = () => {
-    setMenuIsOpen(false);
-  };
-
-  return (
-    <MenuContext.Provider value={{ menuIsOpen, closeMenu, openMenu }}>
-      {children}
-    </MenuContext.Provider>
-  );
-};
