+++ src/ui/hooks/index.ts
@@ -2,3 +2,4 @@ export * from "./useConnections";
 export * from "./useDarkMode";
 export * from "./useModal";
 export * from "./useMenu";
+export * from "./useTitleBar";
